refactor yeti

+++ utils/yetiroutefunctions.py
@@ -255,8 +255,8 @@ def sendTransaction(request, currentroute, nextroute):
         print(v.amo)
         v.amo = "{:.8f}".format(float(v.amo))
         print(v.amo)
-        print(int(v.amo))
-        if int(v.amo) <= 0:
+        print(float(v.amo))
+        if float(v.amo) <= 0:
             v.error = "Amount("+str(v.amo)+") is too small to account for the fee. Try sending a larger amount."
         response = handleResponse('~/yeticold/bitcoin/bin/bitcoin-cli -rpcwallet=yetiwarmpriv createrawtransaction \'[{ "txid": "'+v.sourceaddress['txid']+'", "vout": '+str(v.sourceaddress['vout'])+'}]\' \'[{"'+v.receipentaddress+'" : '+str(v.amo)+'}]\'')
         transonehex = response[:-1]
